==> crop-seq-pipeline(post-QC)/src/glmGamPoi_NE/01.NoBiological_Replicate.ipynb <==
## Step 1: Load Data and Libraries

## Load Libraries
import scanpy as sc
import pandas as pd
import numpy as np
import anndata as ad

## Load sublibrary 1
adata = sc.read_h5ad("/gstore/scratch/u/ghaffars/Dataset/sublib1/raw_qc.h5ad")
adata

adata.obs["gene_symbol"].nunique()

## Step 2: Obtain Guides Labeled as Positive Controls in Sublibrary 1

## Obtain list of positive controls
positive_controls = adata[adata.obs["Biological_replicate"] == "POSITIVE_CONTROL",]

positive_controls

## Examine the Positive Control Targets
table_pos_cont = pd.crosstab(index = positive_controls.obs["gene_symbol"], columns = "count")
table_pos_cont 

It was discussed that those targets with only a few counts in the POSITIVE CONTROL data are likely mislabeled. The true electroporation positive control guides are those with a high frequency. So we will separate only those out. 

## Step 3: Cleaning the Positive Control to remove the samples that are likely mislabled

## Targets with counts greater than 3
print(table_pos_cont['count'][table_pos_cont['count'] > 3])

## Save names of target
pos_control_targets = table_pos_cont['count'][table_pos_cont['count'] > 3].index.tolist()



pos_control_targets

NTC's are likely not a positive control, so we will keep them as well 

## Remove NTC's
try:
    pos_control_targets.remove("NTC")
except ValueError:
    pass

## Keep only positive controls with those targets
positive_controls_clean = positive_controls[positive_controls.obs["gene_symbol"].isin(pos_control_targets),]


positive_controls_clean

## Step 4: Remove positive control samples and save data

## List of all samples
all_samples = adata.obs.index

## List of positive control samples
pos_control_samples = positive_controls_clean.obs.index

pos_control_samples

## Remove positive controls from data
adata_no_pos_controls = adata[~all_samples.isin(pos_control_samples),]


adata_no_pos_controls

## Currently missing .X object, so must make new ann data object
adata_no_pos_controls_out = ad.AnnData(adata_no_pos_controls.layers["counts"])

## Add row and column names
adata_no_pos_controls_out.obs_names = adata_no_pos_controls.obs_names
adata_no_pos_controls_out.var_names = adata_no_pos_controls.var_names

## Add sample info and gene info
adata_no_pos_controls_out.obs = adata_no_pos_controls.obs
adata_no_pos_controls_out.var = adata_no_pos_controls.var

## Store as a copy just in case to make future calculations easier
adata_no_pos_controls_out_final = adata_no_pos_controls_out.copy()

adata_no_pos_controls_out_final.layers["counts"] = adata_no_pos_controls_out_final.X

## Save h5ad object 
adata_no_pos_controls_out_final.write_h5ad("/gstore/scratch/u/ghaffars/glmGamPoi/sublib1_bdev/data/remove_pos_cont_counts_obs_var.h5ad")

## Save sample data
adata_no_pos_controls_out_final.obs.to_csv("/gstore/scratch/u/ghaffars/glmGamPoi/sublib1_bdev/data/remove_pos_cont_sample_data.csv")

## Save gene data
adata_no_pos_controls_out_final.var.to_csv("/gstore/scratch/u/ghaffars/glmGamPoi/sublib1_bdev/data/remove_pos_cont_gene_data.csv")
